Refactor permissions folder structure (WIP)
