+++ front/app/containers/Admin/projects/project/permissions/containers/Phase/index.tsx
@@ -8,8 +8,8 @@ import {
 } from '@citizenlab/cl2-component-library';
 import { useParams, useSearchParams } from 'react-router-dom';
 import { FormattedMessage } from 'utils/cl-intl';
-import PhasePermissions from './granular_permissions/containers/Granular/PhasePermissions';
-import messages from './messages';
+import PhasePermissions from '../Project/Granular/PhasePermissions';
+import messages from '../../messages';
 import { updateSearchParams } from 'utils/cl-router/updateSearchParams';
 import { removeSearchParams } from 'utils/cl-router/removeSearchParams';
 import useFeatureFlag from 'hooks/useFeatureFlag';
