+++ front/app/containers/Admin/projects/project/permissions/containers/Project/Granular/AddFieldScreen.tsx
@@ -31,8 +31,8 @@ import { isNilOrError } from 'utils/helperUtils';
 import validateMultiloc from 'utils/yup/validateMultilocForEveryLocale';
 import validateOneOptionForMultiSelect from 'utils/yup/validateOneOptionForMultiSelect';
 
-import messages from '../../../containers/Granular/messages';
-import { getLabelForInputType } from '../../../containers/Granular/utils';
+import messages from './messages';
+import { getLabelForInputType } from './utils';
 
 type AddFieldScreenProps = {
   setShowAddFieldPage: (show: boolean) => void;
