+++ front/app/containers/Admin/projects/project/permissions/containers/Project/ProjectVisibility/GroupAvatar.tsx
