+++ front/app/containers/Admin/projects/routes.tsx
@@ -34,9 +34,11 @@ const AdminPhaseNewAndEdit = lazy(() => import('./project/timeline/edit'));
 const AdminProjectEvents = lazy(() => import('./project/events'));
 const AdminProjectEventsEdit = lazy(() => import('./project/events/edit'));
 const AdminProjectPermissions = lazy(
-  () => import('./project/permissions/Project')
+  () => import('./project/permissions/containers/Project')
 );
-const AdminPhasePermissions = lazy(() => import('./project/permissions/Phase'));
+const AdminPhasePermissions = lazy(
+  () => import('./project/permissions/containers/Phase')
+);
 const AdminProjectSurveyResults = lazy(() => import('./project/surveyResults'));
 const AdminProjectPoll = lazy(() => import('./project/poll'));
 const AdminProjectsSurvey = lazy(() => import('./project/nativeSurvey'));
